backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# OCR and PDF processing
pytesseract==0.3.10
pdf2image==1.16.3
PyPDF2==3.0.1
pdfminer.six==20221105
Pillow==10.1.0

# NLP and embeddings
spacy==3.7.2
sentence-transformers>=3.0.0
transformers>=4.40.0

# Vector database
faiss-cpu==1.7.4
numpy==1.24.3

# LLM integration
openai==1.3.5
langchain==0.0.335
langchain-community==0.0.1

# Data processing
pandas==2.1.3
python-dateutil==2.8.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
aiofiles==23.2.1

# Database
pymongo==4.6.0
dnspython==2.4.2

# HTTP requests
requests==2.31.0


# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
scikit-learn==1.3.2
